MappingConfiguration: Fix EFCoreMappingConfiguration ctor.
This commit contains a fix for the EFCoreMappingConfiguration
constructor, ensuring that *it* handles name mapping, and not the base
class constructor. This ensures that the `_navigation` suffixes are
already truncated from the name mappings by the time the provided
dictionary is used for mapping UCAST names to object properties.

An extra mapping and test class were added to the unit tests to ensure
coverage of this code path.

Signed-off-by: Philip Conrad <philip@chariot-chaser.net>